Add the -via arguments to config file and GUI for the viewer
vncviewer already includes a `-via <ssh_host>` command line option to
specify that an SSH tunnel to `<ssh_host>` should be established and
VNC traffic routed through the tunnel.  However, that option was not
available either in the `*.tigervnc` setup files or via the GUI.

This adds `UseSSH` (boolean) and `via` (string) parameters to the config
file and exposes them to the VNC viewer GUI under the options dialog.