Only try to establish an SSH tunnel if the host is not blank.
Check that the string holding the SSH host name is not blank.  This
prevents a host like "   " from causing confusing errors if someone
manages to get that entered.  If the string is empty (zero length)
or only has whitespace (not empty but blank) then do not try
to start an SSH tunnel.