Add the -via arguments to config file and GUI for the viewer
vncviewer already includes a `-via <ssh_host>` command line option to
specify that an SSH tunnel to `<ssh_host>` should be established and
VNC traffic routed through the tunnel.  However, that option was not
available either in the `*.tigervnc` setup files or via the GUI.

This adds a `via` (string) parameter to the config file and exposes
it to the VNC viewer GUI under the options dialog.

In addition, when establishing a tunnel and even before deciding to
establish a tunnel, strip leading and trailing whitespace because
`ssh "    my.host.com   "` will indeed fail and if someone manages
to get a blank value (only whitespace) into the GUI or the config file
it can produce confusing results.